Potential flaw in the gradient based addition code
We want to prevent getting stuck in(read: exit) local minima with it, but I placed an assertion for non-zero value for mx.

Simply removing this asserting will not help, even if we don't actually need normalization; if it is completely stuck in the local minima(mx=0), I think adding a new gate with parameter=0 will keep it there. Options I see:
- Just remove normalization with mx, there's really no need for it. I think it will remain in the local minima, but confirm it.
- Every new gate has random parameter.
- If mx is not 0, continue to use a 0 parameter gate, if mx is 0, add the new gate with a random parameter instead of zero.

Even for the last two options, a problem remains: how to determine which gate to add when mx is 0?

Alternate: until mx is not 0, use gradient based gate addition, when it is 0, add a new layer? Maybe with random parameters